mv dbconn to pkg/dbconn
With the introduction of the management console, the frontend will no longer be
the only service which talks to the database. Therefor, sharing the logic to
connect to and migrate the database makes sense. This change does just that.

It was also sharable before, but it would have meant importing
`cmd/frontend/db/dbconn` from the management console which does not make sense
logically as they are separate services.

Migration logic still lives in pkg/dbconn, because we only have one database
(not multiple).

Helps #965