mv cmd/frontend/db/migrations -> migrations/
Previously the generated contents of the migrations/ folder was stored
in a subpackage of the frontend. This no longer makes sense with pkg/dbconn
importing it.

Additionally, this is something that confuses me every time I add a migration; i.e.
"what directory do I generate and why is it not the same as the directory I modified?!"